pyproject.toml
[project]
name = "mimir-agent"
version = "0.1.0"
description = "Product knowledge agent built on Norns durable runtime."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "norns-sdk>=0.2.0",
    "PyGithub>=2.1",
    "httpx>=0.27",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "slack-bolt>=1.18",
    "watchfiles>=1.0",  # dev: auto-restart on file changes
]

[project.scripts]
mimir-agent = "mimir_agent.main:main"
mimir-worker = "mimir_agent.worker:main"
mimir-slack = "mimir_agent.slack_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
